DAOS-10244 test: Create a new file, agent_failure_isolate.py
Skip-unit-tests: true
Test-tag: agent_failure
